filter, sort, and reverse transforms
